+++ src/ska_mid_dish_manager/models/fanned_out_command.py
@@ -1,6 +1,7 @@
 """Abstracts all the logic for executing a command on a device."""
 
 import enum
+import json
 import logging
 import time
 from typing import Any, Callable, Optional
@@ -60,11 +61,11 @@ class FannedOutCommand:
         self.command_argument = command_argument
         self.timeout_s = timeout_s
         self.start_time: float = 0.0
-        self.cmd_message = None
+        self.executed_cmd_message = None
         self._status = FannedOutCommandStatus.PENDING
         self._task_finish_reported = False
         self._progress_callback = progress_callback
-        self.cmd_response = ""
+        self.executed_cmd_response = ""
         self.component_state = component_state
         self.awaited_component_state = awaited_component_state
         self.awaited_update_reports = {attr: False for attr in awaited_component_state.keys()}
@@ -72,7 +73,7 @@ class FannedOutCommand:
     def execute(self, task_callback: Callable) -> None:
         """Execute the fanned out command."""
         self.logger.debug(f"Executing {self.command_name} with arg {self.command_argument}")
-        self._status = FannedOutCommandStatus.RUNNING
+        self._status = FannedOutCommandStatus.IN_PROGRESS
         self.start_time = time.time()
 
         try:
@@ -80,11 +81,18 @@ class FannedOutCommand:
             assert len(res) == 2, (
                 f"FannedOutCommand 'command' Callable expects a response of len 2, but got '{res}'"
             )
-            self.cmd_response, self.cmd_message = res
+            self.executed_cmd_response, self.executed_cmd_message = res
+
+            if self.awaited_component_state is not None:
+                awaited_attributes = list(self.awaited_component_state.keys())
+                awaited_values = list(self.awaited_component_state.values())
+                report_awaited_attributes(
+                    self._progress_callback, awaited_attributes, awaited_values, self.device
+                )
         except RuntimeError as e:
             self.logger.error(f"FannedOutCommand '{self.command_name}' failed to execute: {e}")
             self._status = FannedOutCommandStatus.FAILED
-            self.cmd_response = f"{e.args[0]}"
+            self.executed_cmd_response = f"{e.args[0]}"
 
     @property
     def status(self) -> FannedOutCommandStatus:
@@ -94,7 +102,12 @@ class FannedOutCommand:
     @property
     def failed(self) -> bool:
         """Check if the fanned out command has failed."""
-        return self.status in (FannedOutCommandStatus.TIMED_OUT, FannedOutCommandStatus.FAILED)
+        return self.status in (
+            FannedOutCommandStatus.TIMED_OUT,
+            FannedOutCommandStatus.FAILED,
+            FannedOutCommandStatus.ABORTED,
+            FannedOutCommandStatus.REJECTED,
+        )
 
     @property
     def successful(self) -> bool:
@@ -107,28 +120,22 @@ class FannedOutCommand:
         return self.failed or self.successful
 
     def _update_status(self, task_callback: Callable) -> None:
-        if self._status == FannedOutCommandStatus.RUNNING:
-            # timeout
-            if self.timeout_s > 0 and time.time() - self.start_time > self.timeout_s:
-                self._status = FannedOutCommandStatus.TIMED_OUT
+        """Update the status of the command based on component state and timeout checks."""
+        if self._status == FannedOutCommandStatus.IN_PROGRESS:
             # completed
             if check_component_state_matches_awaited(
                 self.component_state, self.awaited_component_state
             ):
                 self._status = FannedOutCommandStatus.COMPLETED
-        if self._status in [FannedOutCommandStatus.FAILED, FannedOutCommandStatus.TIMED_OUT]:
-            report_task_progress(
-                f"{self.device} device {self._status.name.lower().replace('_', ' ')}"
-                f" executing {self.command_name} command",
-                self._progress_callback,
-            )
+            # timeout
+            if self.timeout_s > 0 and time.time() - self.start_time > self.timeout_s:
+                self._status = FannedOutCommandStatus.TIMED_OUT
 
     def report_progress(self, task_callback: Callable) -> None:
         """Report the progress of fanned out command."""
-        self._update_status(task_callback)
         current_comp_state = dict(self.component_state)
 
-        # Awaited component state updates
+        # Report awaited component state updates
         for attr_name, reported_update in self.awaited_update_reports.items():
             if not reported_update and attr_name in self.awaited_component_state:
                 expected_value = self.awaited_component_state[attr_name]
@@ -142,15 +149,37 @@ class FannedOutCommand:
                             self._progress_callback,
                         )
                         self.awaited_update_reports[attr_name] = True
+
+        # Update the commands status
+        self._update_status(task_callback)
+
+        # Report if the command has finished
         if self.finished and not self._task_finish_reported:
             status_name = self.status.name.lower().replace("_", " ")
-            report_task_progress(
-                f"{self.device}.{self.command_name} {status_name}", self._progress_callback
+            cmd_response = (
+                self.executed_cmd_response
+                if self._status
+                in [
+                    FannedOutCommandStatus.FAILED,
+                    FannedOutCommandStatus.REJECTED,
+                    FannedOutCommandStatus.ABORTED,
+                ]
+                else ""
             )
+            if cmd_response:
+                report_task_progress(
+                    f"{self.device}.{self.command_name} {status_name}: {cmd_response}",
+                    self._progress_callback,
+                )
+            else:
+                report_task_progress(
+                    f"{self.device}.{self.command_name} {status_name}",
+                    self._progress_callback,
+                )
             self._task_finish_reported = True
 
 
-class FannedOutSlowCommand(FannedOutCommand):
+class FannedOutTangoCommand(FannedOutCommand):
     def __init__(
         self,
         logger: logging.Logger,
@@ -213,16 +242,179 @@ class FannedOutSlowCommand(FannedOutCommand):
         task_status, msg = self.device_component_manager.execute_command(
             self.command_name, self.command_argument
         )
-        if self.awaited_component_state is not None:
-            awaited_attributes = list(self.awaited_component_state.keys())
-            awaited_values = list(self.awaited_component_state.values())
-            report_awaited_attributes(
-                self._progress_callback, awaited_attributes, awaited_values, self.device
-            )
-
-        if task_status == TaskStatus.FAILED:
+        if task_status in [TaskStatus.FAILED, TaskStatus.REJECTED, TaskStatus.ABORTED]:
             raise RuntimeError(msg)
         return task_status, msg
+
+
+class FannedOutTangoLongRunningCommand(FannedOutTangoCommand):
+    def __init__(
+        self,
+        logger: logging.Logger,
+        device: str,
+        command_name: str,
+        device_component_manager: TangoDeviceComponentManager,
+        command_argument: Any = None,
+        awaited_component_state: dict = {},
+        timeout_s: float = 0,
+        progress_callback: Optional[Callable] = None,
+        is_device_ignored: bool = False,
+    ):
+        """:param logger: Logger instance
+        :type logger: Logger
+        :param device: The name for the device this command is executed on
+        :type device: str
+        :param command_name: The name for the command to be executed
+        :type command_name: str
+        :param device_component_manager: The component manager of the subservient device
+        :type device_component_manager: TangoDeviceComponentManager
+        :param timeout_s: Timeout (in seconds) for the command execution
+        :type timeout_s: float
+        :param command_argument: Argument for the requested command
+        :type command_argument: Any
+        :param awaited_component_state: The component state containing the attributes and values to
+            wait for.
+        :type awaited_component_state: dict
+        :param progress_callback: Optional callback to report progress updates.
+        :type progress_callback: Callable
+        :param is_device_ignored: Toggle to ignore fanning out of command if the device is ignored.
+        :type is_device_ignored: bool
+        """
+        self.is_lrc_finished = False
+        super().__init__(
+            logger=logger,
+            device=device,
+            command_name=f"{command_name}",
+            device_component_manager=device_component_manager,
+            command_argument=command_argument,
+            awaited_component_state=awaited_component_state,
+            timeout_s=timeout_s,
+            progress_callback=progress_callback,
+            is_device_ignored=is_device_ignored,
+        )
+
+    def _execute_tango_command(self) -> tuple:
+        """Fan out the respective command to the device and handle task status response."""
+        task_status, msg = super()._execute_tango_command()
+
+        # If the command completed immediately then it won't appear in the LRC attributes. Mark
+        # the lrc as complete, the component state check will be used to complete the command.
+        if task_status in [TaskStatus.COMPLETED, TaskStatus.REJECTED, TaskStatus.ABORTED]:
+            self.is_lrc_finished = True
+
+        return task_status, msg
+
+    def _is_command_in_lrc_queued(self) -> bool:
+        """Check if the long running command is in the lrcQueue attribute."""
+        lrc_queue = self.device_component_manager.read_attribute_value("lrcqueue", log_read=False)
+        if not isinstance(lrc_queue, tuple):
+            self.logger.error(
+                "lrcQueue value is not a tuple, got %s: %s", type(lrc_queue), lrc_queue
+            )
+            return False
+
+        for queued_cmd in lrc_queue:
+            try:
+                queued_cmd_dict = json.loads(queued_cmd)
+            except json.JSONDecodeError:
+                self.logger.exception("Invalid json value for lrcQueue")
+                continue
+
+            if queued_cmd_dict.get("uid") == self.executed_cmd_message:
+                return True
+        return False
+
+    def _is_command_in_lrc_executing(self) -> bool:
+        """Check if the long running command is in the lrcExecuting attribute."""
+        lrc_executing = self.device_component_manager.read_attribute_value(
+            "lrcexecuting", log_read=False
+        )
+        if not isinstance(lrc_executing, tuple):
+            self.logger.error(
+                "lrcExecuting value is not a tuple, got %s: %s", type(lrc_executing), lrc_executing
+            )
+            return False
+
+        for executing_cmd in lrc_executing:
+            try:
+                executing_cmd_dict = json.loads(executing_cmd)
+            except json.JSONDecodeError:
+                self.logger.exception("Invalid json value for lrcExecuting")
+                continue
+
+            if executing_cmd_dict.get("uid") == self.executed_cmd_message:
+                return True
+        return False
+
+    def _get_command_lrc_finished_dict(self) -> Optional[dict]:
+        """Get the lrcFinished dict for the long running command."""
+        lrc_finished = self.device_component_manager.read_attribute_value(
+            "lrcfinished", log_read=False
+        )
+        if not isinstance(lrc_finished, tuple):
+            self.logger.error(
+                "lrcFinished value is not a tuple, got %s: %s", type(lrc_finished), lrc_finished
+            )
+            return None
+
+        for finished_cmd in lrc_finished:
+            try:
+                finished_cmd_dict = json.loads(finished_cmd)
+            except json.JSONDecodeError:
+                self.logger.exception("Invalid json value for lrcFinished")
+                return None
+
+            if finished_cmd_dict.get("uid") == self.executed_cmd_message:
+                return finished_cmd_dict
+        return None
+
+    def _update_status(self, task_callback: Callable) -> None:
+        """Update the status of the fanned out command based on the LRC status and component state.
+
+        Requires both the LRC to have completed and the component states to match to complete.
+        """
+        if self._status in [FannedOutCommandStatus.QUEUED, FannedOutCommandStatus.IN_PROGRESS]:
+            # If the LRC has not yet been reported in lrcFinished
+            if not self.is_lrc_finished:
+                lrc_finished_dict = self._get_command_lrc_finished_dict()
+
+                if lrc_finished_dict:
+                    lrc_result = lrc_finished_dict["result"]
+                    lrc_status = lrc_finished_dict["status"]
+
+                    self.executed_cmd_response = lrc_result
+
+                    if lrc_status == TaskStatus.COMPLETED.name:
+                        # Don't mark it as completed yet, still need to check component state
+                        self.is_lrc_finished = True
+                    elif lrc_status == TaskStatus.ABORTED.name:
+                        self._status = FannedOutCommandStatus.ABORTED
+                        return
+                    elif lrc_status == TaskStatus.REJECTED.name:
+                        self._status = FannedOutCommandStatus.REJECTED
+                        return
+                    elif lrc_status == TaskStatus.FAILED.name:
+                        self._status = FannedOutCommandStatus.FAILED
+                        return
+                elif self._is_command_in_lrc_executing():
+                    self._status = FannedOutCommandStatus.IN_PROGRESS
+                elif self._is_command_in_lrc_queued():
+                    self._status = FannedOutCommandStatus.QUEUED
+
+            # Final component state check
+            if self.is_lrc_finished:
+                component_ready = check_component_state_matches_awaited(
+                    self.component_state,
+                    self.awaited_component_state,
+                )
+
+                if component_ready:
+                    self._status = FannedOutCommandStatus.COMPLETED
+                    return
+
+            # timeout
+            if self.timeout_s > 0 and time.time() - self.start_time > self.timeout_s:
+                self._status = FannedOutCommandStatus.TIMED_OUT
 
 
 class DishManagerCMMethod(FannedOutCommand):
@@ -263,17 +455,17 @@ class DishManagerCMMethod(FannedOutCommand):
                 f"and kwargs. {self.command_kwargs}"
             )
         )
-        self._status = FannedOutCommandStatus.RUNNING
+        self._status = FannedOutCommandStatus.IN_PROGRESS
         self.start_time = time.time()
         try:
             res = self.command(*self.command_args, **self.command_kwargs)
             self.logger.debug(f"Result: {res}")
-            self.cmd_response = res
+            self.executed_cmd_response = res
             self._status = FannedOutCommandStatus.COMPLETED
         except Exception as e:
             self.logger.exception(f"FannedOutCommand '{self.command_name}' failed to execute: {e}")
             self._status = FannedOutCommandStatus.FAILED
-            self.cmd_response = f"{e}"
+            self.executed_cmd_response = f"{e}"
 
 
 class DishManagerCMMethodCallBack(FannedOutCommand):
@@ -315,11 +507,11 @@ class DishManagerCMMethodCallBack(FannedOutCommand):
             if status in (TaskStatus.FAILED, TaskStatus.ABORTED, TaskStatus.NOT_FOUND):
                 self._status = FannedOutCommandStatus.FAILED
             if status in (TaskStatus.QUEUED, TaskStatus.STAGING, TaskStatus.IN_PROGRESS):
-                self._status = FannedOutCommandStatus.RUNNING
+                self._status = FannedOutCommandStatus.IN_PROGRESS
 
     def execute(self, task_callback) -> None:
         """Execute the command."""
-        self._status = FannedOutCommandStatus.RUNNING
+        self._status = FannedOutCommandStatus.IN_PROGRESS
         self.start_time = time.time()
         self.command_args = list(self.command_args)
         self.command_args.insert(0, self._task_callback)
@@ -332,11 +524,11 @@ class DishManagerCMMethodCallBack(FannedOutCommand):
             )
             res = self.command(*self.command_args, **self.command_kwargs)
             self.logger.debug(f"Result: {res}")
-            self.cmd_response = res
+            self.executed_cmd_response = res
         except Exception as e:
             self.logger.exception(f"FannedOutCommand '{self.command_name}' failed to execute: {e}")
             self._status = FannedOutCommandStatus.FAILED
-            self.cmd_response = f"{e}"
+            self.executed_cmd_response = f"{e}"
 
 
 class DishManagerCMMethodResultCode(FannedOutCommand):
@@ -370,7 +562,7 @@ class DishManagerCMMethodResultCode(FannedOutCommand):
 
     def execute(self, task_callback) -> None:
         """Execute the command."""
-        self._status = FannedOutCommandStatus.RUNNING
+        self._status = FannedOutCommandStatus.IN_PROGRESS
         self.start_time = time.time()
         try:
             self.logger.debug(
@@ -381,7 +573,7 @@ class DishManagerCMMethodResultCode(FannedOutCommand):
             )
             result_code, message = self.command(*self.command_args, **self.command_kwargs)
             self.logger.debug(f"Result: {result_code}, Message: {message}")
-            self.cmd_response = result_code
+            self.executed_cmd_response = result_code
             # For DishManagerCMMethodResultCode, we expect an immediate response.
             # Any response that gets queued/aborted/etc is considered failed.
             # In those cases use another Action.
@@ -392,4 +584,4 @@ class DishManagerCMMethodResultCode(FannedOutCommand):
         except Exception as e:
             self.logger.exception(f"FannedOutCommand '{self.command_name}' failed to execute: {e}")
             self._status = FannedOutCommandStatus.FAILED
-            self.cmd_response = f"{e}"
+            self.executed_cmd_response = f"{e}"