+++ tests/unit/business_logic/test_command_actions_ignoring_devices.py
@@ -1,11 +1,12 @@
 """Tests for ignoring devices in command Action."""
 
+import json
 import logging
 from threading import Event
 from unittest import mock
 
 import pytest
-from ska_control_model import AdminMode
+from ska_control_model import AdminMode, TaskStatus
 
 from ska_mid_dish_manager.models.command_actions import SetStandbyLPModeAction
 from ska_mid_dish_manager.models.dish_enums import (
@@ -32,7 +33,18 @@ class TestCommandActionsIgnoringDevices:
                     "operatingmode": DSOperatingMode.STANDBY,
                     "powerstate": DSPowerState.LOW_POWER,
                 },
-                execute_command=mock.MagicMock(return_value=(None, None)),
+                execute_command=mock.MagicMock(return_value=(None, "command_id_123")),
+                read_attribute_value=mock.MagicMock(
+                    return_value=(
+                        json.dumps(
+                            {
+                                "uid": "command_id_123",
+                                "result": "command_id_123",
+                                "status": TaskStatus.COMPLETED.name,
+                            }
+                        ),
+                    )
+                ),
             ),
             "SPF": mock.MagicMock(
                 _component_state={"operatingmode": SPFOperatingMode.STANDBY_LP},