+++ tests/unit/interface/tango/test_abort.py
@@ -77,9 +77,6 @@ def test_abort_during_dish_movement(dish_manager_resources, event_store_class, p
     spf_cm = dish_manager_cm.sub_component_managers["SPF"]
     spfrx_cm = dish_manager_cm.sub_component_managers["SPFRX"]
 
-    # update the execute_command mock to return IN_PROGRESS and a timestamp
-    ds_cm.execute_command = Mock(return_value=(TaskStatus.IN_PROGRESS, 1234567890.0))
-
     dish_mode_event_store = event_store_class()
     status_event_store = event_store_class()
     cmds_in_queue_store = event_store_class()
@@ -153,9 +150,6 @@ def test_abort_sends_lrc_update_after_abort_sequence_finishes(
     ds_cm = dish_manager_cm.sub_component_managers["DS"]
     spf_cm = dish_manager_cm.sub_component_managers["SPF"]
     spfrx_cm = dish_manager_cm.sub_component_managers["SPFRX"]
-
-    # update the execute_command mock to return IN_PROGRESS and a timestamp
-    ds_cm.execute_command = Mock(return_value=(TaskStatus.IN_PROGRESS, 1234567890.0))
 
     dish_mode_event_store = event_store_class()
     lrc_result_event_store = event_store_class()